requirements.txt
sentence-transformers
langchain
pypdf
langchain-community
langchain_huggingface
faiss-cpu
sumy
streamlit
sentencepiece
